+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 0.2.0 (in progress)
 -------------------
 
+- Added payment status transition endpoint
+- Added get bank account status transition endpoint
 - Added list program account receipt endpoint
 - Added list user receipt endpoint
 - Added list prepaid card receipt endpoint